Add test vec_add

+++ tests/matrix_transpose/matrix_transpose.cuh
@@ -1,6 +1,6 @@
 struct MatrixTransposeArgs {
     float* idata, *odata;
-    int width;
+    unsigned long width;
 };
 
 struct MatrixTransposeContext{
